requirements.txt
jupyterlab>=4.0.0
inspect-ai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
numpy>=1.24.0
seaborn>=0.12.0
ipykernel>=6.25.0
python-dotenv>=1.0.0
